@@ -24,6 +24,7 @@
   },
   "homepage": "https://github.com/OpusCapita/onboarding/blob/master/README.md",
   "dependencies": {
+    "assets-webpack-plugin": "3.5.1",
     "async": "2.3.0",
     "babel-runtime": "6.23.0",
     "bluebird": "3.5.0",
@@ -33,6 +34,7 @@
     "epilogue": "0.7.1",
     "express": "4.15.2",
     "express-handlebars": "3.0.0",
+    "extract-text-webpack-plugin": "2.1.0",
     "jquery": "2.1.4",
     "lodash": "4.17.4",
     "moment": "2.18.1",
@@ -42,14 +44,15 @@
     "nodemon": "1.11.0",
     "ocbesbn-config": "^2.2.5",
     "ocbesbn-db-init": "^1.1.5",
-    "ocbesbn-web-init": "^1.0.3",
-    "opuscapita-i18n": "1.0.11",
+    "ocbesbn-react-components": "^0.1.1",
+    "ocbesbn-web-init": "^1.1.1",
+    "opuscapita-i18n": "1.0.12",
     "papaparse": "4.2.0",
     "polyfill-promise": "4.0.1",
     "react": "15.4.2",
     "react-addons-test-utils": "15.4.2",
     "react-addons-transition-group": "15.4.2",
-    "react-bootstrap": "0.30.8",
+    "react-bootstrap": "0.30.9",
     "react-bootstrap-table": "2.5.7",
     "react-cookie": "1.0.5",
     "react-dom": "15.4.2",
@@ -57,15 +60,15 @@
     "react-dropzone-component": "1.4.1",
     "react-intl": "2.2.3",
     "react-notification-system": "0.2.13",
-    "react-redux": "5.0.3",
+    "react-redux": "5.0.4",
     "react-router": "3.0.2",
     "react-router-redux": "4.0.8",
     "react-spinkit": "1.1.11",
     "react-utils-helper": "0.1.4",
-    "recharts": "0.22.0",
+    "recharts": "0.22.2",
     "redis": "2.7.1",
     "redux": "3.6.0",
-    "redux-form": "6.6.1",
+    "redux-form": "6.6.3",
     "redux-thunk": "2.2.0",
     "rimraf": "2.6.1",
     "sequelize": "3.30.4",
@@ -76,12 +79,14 @@
     "umzug": "1.11.0",
     "uuid": "3.0.1",
     "validate.js": "0.11.1",
-    "xlsx": "0.9.9"
+    "webpack-config": "7.0.0",
+    "webpack-md5-hash": "0.0.5",
+    "xlsx": "0.9.11"
   },
   "devDependencies": {
-    "babel-cli": "6.24.0",
-    "babel-core": "6.24.0",
-    "babel-eslint": "7.2.1",
+    "babel-cli": "6.24.1",
+    "babel-core": "6.24.1",
+    "babel-eslint": "7.2.2",
     "babel-jest": "19.0.0",
     "babel-loader": "6.4.1",
     "babel-plugin-istanbul": "4.1.1",
@@ -90,11 +95,11 @@
     "babel-plugin-transform-runtime": "6.23.0",
     "babel-plugin-webpack-loaders": "0.9.0",
     "babel-polyfill": "6.23.0",
-    "babel-preset-es2015": "6.24.0",
+    "babel-preset-es2015": "6.24.1",
     "babel-preset-jest": "19.0.0",
-    "babel-preset-react": "6.23.0",
-    "babel-preset-stage-0": "6.22.0",
-    "babel-register": "6.24.0",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-0": "6.24.1",
+    "babel-register": "6.24.1",
     "chai": "3.5.0",
     "command-line-args": "4.0.2",
     "css-loader": "0.28.0",
@@ -110,11 +115,9 @@
     "nodemon": "1.11.0",
     "npm-run-all": "4.0.2",
     "nyc": "10.2.0",
-    "raml-to-markdown": "1.1.0",
+    "raml-to-markdown": "1.1.1",
     "sequelize-to-markdown": "1.1.12",
     "style-loader": "0.16.1",
-    "webpack": "2.3.3",
-    "webpack-dev-middleware": "1.10.1",
-    "webpack-hot-middleware": "2.18.0"
+    "webpack": "2.4.1"
   }
 }
