@@ -1,18 +1,12 @@
 import React from 'react';
 import { CampaignsApplication } from './containers'
 import { render } from 'react-dom';
-import campaignRoutes from './routes';
-import campaignsReducer from './reducers';
 
-module.exports = {
+export default {
   renderCampaignEditor: function(domElement, props) {
     render(
-            <CampaignsApplication {...props}/>,
-            domElement
-        );
-  },
-
-  campaignsRoutes: campaignRoutes,
-
-  campaignsReducer: campaignsReducer,
+      <CampaignsApplication {...props}/>,
+      domElement
+    );
+  }
 };
